+++ src/main/java/de/is24/deadcode4j/analyzer/ServletContainerInitializerAnalyzer.java
@@ -1,12 +1,15 @@
 package de.is24.deadcode4j.analyzer;
 
+import com.google.common.collect.Maps;
 import de.is24.deadcode4j.Analyzer;
 import de.is24.deadcode4j.CodeContext;
+import de.is24.deadcode4j.IntermediateResult;
 import de.is24.deadcode4j.Module;
 import org.xml.sax.Attributes;
 import org.xml.sax.helpers.DefaultHandler;
 
 import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
 import java.io.File;
 import java.util.Map;
 
@@ -93,12 +96,19 @@ public class ServletContainerInitializerAnalyzer extends AnalyzerAdapter {
         private boolean metadataComplete = false;
 
         private ServletContainerInitializerCodeContext(Module module) {
-            super(module);
+            super(module, Maps.<Object, IntermediateResult>newHashMap());
         }
 
+        @Nonnull
         @Override
         public Map<Object, Object> getCache() {
             return this.originalContext.getCache();
+        }
+
+        @Nullable
+        @Override
+        public IntermediateResult getIntermediateResult(@Nonnull Object key) {
+            return this.originalContext.getIntermediateResult(key);
         }
 
         @Override