+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 
 ## Internal
 * calculate class path for each project & pass along to analyzers
+    * analyze each module with its own context
+    * establish possibility to pass on intermediate results to depending modules
 
 ###[code changes](https://github.com/ImmobilienScout24/deadcode4j/compare/deadcode4j-maven-plugin-1.5...master)